requirements.txt
picamera
numpy
RPi.GPIO